autocrypto_social_bot/requirements.txt
selenium==4.11.2
openai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
undetected-chromedriver==3.5.3
webdriver-manager==4.0.0
setuptools>=65.5.1